Reorganized file importer tests to file_importer_test and added a module to reuse setup and teardown from file_submission_test

+++ data_engineering/test/test_helper.rb
@@ -1,5 +1,8 @@
 ENV["RAILS_ENV"] = "test"
 require File.expand_path('../../config/environment', __FILE__)
+# probably a better way to require or autoload this
+# but I was running short on patience
+require File.expand_path('test/unit/helpers/file_submission_test_helper')
 require 'rails/test_help'
 
 class ActiveSupport::TestCase
@@ -8,7 +11,7 @@ class ActiveSupport::TestCase
   # get rid of the file so we don't get a futre
   # false positive on a file that exists
   def remove_test_upload_file(file_submission)
-    if File.exists?(file_submission.data_file.path)
+    if !file_submission.data_file.path.nil? && File.exists?(file_submission.data_file.path)
       require 'fileutils'
       FileUtils.rm_rf(File.join(Rails.root, 'test/fixtures/paperclip_files/file_submissions'))
     end
